src/django-credentials/requirements.txt
Django==4.0.5
gunicorn
hvac
psycopg2-binary
django-cors-headers
djangorestframework
djangorestframework-simplejwt
Pillow
django-redis==5.4.0
celery
django-celery-beat
channels
channels-redis
daphne